requirements.txt
anyio>=4.4.0
langchain
zipp>=3.19.1
validators==0.34.0
tqdm==4.67.1
ipwhois==1.3.0
fake_useragent
requests>=2.32.3
openai>=0.27.0
beautifulsoup4==4.13.3
lxml>=4.6.3
colorama==0.4.6
google-search-results==2.4.2
httpx==0.28.1
tenacity==9.0.0
requests_html==0.10.0
aiohttp==3.11.12
urllib3==2.3.0
asyncio==3.4.3
lxml_html_clean
html_clean
dnspython>=2.6.1
aiohttp-socks==0.10.1
python-socks[asyncio]>=2.4.3
anthropic
mistralai
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
